chore: rename folder plugins -> markmap-plugins
