+++ src/markmap-plugins/checkbox.ts
